Fix env_universal_common.cpp filesize comparisons.
* if (result == ULLONG_MAX) is always false, likely a typo as
  result is unsigned long, and the comment says ULONG_MAX.

* use off_t instead of size_t for file size where it can mismatch
  st_size's type in stat.h

+++ src/env_universal_common.cpp
@@ -1023,7 +1023,7 @@ class universal_notifier_shmem_poller_t : public universal_notifier_t {
         }
 
         // Get the size.
-        size_t size = 0;
+        off_t size = 0;
         if (!errored) {
             struct stat buf = {};
             if (fstat(fd, &buf) < 0) {
@@ -1352,7 +1352,7 @@ class universal_notifier_named_pipe_t : public universal_notifier_t {
         // Now return the smaller of the two values. If we get ULONG_MAX, it means there's no more
         // need to poll; in that case return 0.
         unsigned long result = mini(readback_delay, polling_delay);
-        if (result == ULLONG_MAX) {
+        if (result == ULONG_MAX) {
             result = 0;
         }
         return result;
